requirements.txt
django
django-heroku
gunicorn
djangorestframework
markdown
django-filter
psycopg2-binary
dj-database-url
django-rest-auth
uuid